backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
websockets>=9.0,<11
alpaca-trade-api>=3.1.1
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0

